+++ include/cinterop/timeseries_interop.hpp
@@ -60,7 +60,7 @@ namespace cinterop
 		template<>
 		inline void dispose_of<time_series_dimensions_description>(time_series_dimensions_description& d)
 		{
-			for (size_t i = 0; i < d.num_dimensions; i++)
+			for (int i = 0; i < d.num_dimensions; i++)
 			{
 				auto p = d.dimensions[i];
 				if (p.dimension_type != nullptr)
